Modularize code following the MVC pattern up to the book selection
I broke down the different classes in separate files, changed main.js to
controller.js and created a constants module that simplifies code that
won't change and that's a bit messy.

I already coded the book selection event, now I need to create a class
for the Library and show the selected book.